+++ Factory/IoC.cs
@@ -5,7 +5,7 @@ using Exceptions.Commands;
 
 #endregion
 
-namespace Fabric
+namespace Factory
 {
     public class IoC : IResolvable
     {
